rename files to pass test and minor changes
