+++ app.js
@@ -8,8 +8,8 @@ app.get("/", (req,res) => {
 const locationsController = require("./controllers/locations.controller.js")
 app.use("/locations", locationsController)
 
-const personsController = require("./controllers/people.controller.js")
-app.use("/people", personsController)
+const personsController = require("./controllers/persons.controller.js")
+app.use("/persons", personsController)
 
 const plansController = require("./controllers/plans.controller.js")
 app.use("/plans", plansController)
@@ -17,11 +17,11 @@ app.use("/plans", plansController)
 const machinesController = require("./controllers/machines.controller.js")
 app.use("/machines", machinesController)
 
-const specialEventsController = require("./controllers/specialEvents.controller.js")
+const specialEventsController = require("./controllers/special-events.controller.js")
 app.use("/special-events", specialEventsController)
 
 app.get("*", (req,res) => {
-    res.json({error: "Sorry, no page found!"})
+    res.status(404).json({error: "Sorry, no page found!"})
 })
 
 module.exports = app
