+++ controllers/plans.controller.js
@@ -1,6 +1,6 @@
 const express = require("express")
 const plans = express.Router()
-const plansArray = require("../models/plans.model.js")
+const plansArray = require("../models/plan.model.js")
 
 plans.get("/", (req,res) => {
     res.json(plansArray)
